refactor(ui): split out joining code utils into seperate file for easier reuse
