+++ ui/src/AttachmentsList.svelte
@@ -2,7 +2,7 @@
   import "@shoelace-style/shoelace/dist/components/skeleton/skeleton.js";
   import { createEventDispatcher, getContext } from "svelte";
   import type { KanDoStore } from "./stores/kando";
-  import type {  WALUrl } from "./util";
+  import type {  WALUrl } from "./utils/util";
   import { weaveUrlToWAL } from "@lightningrodlabs/we-applet";
   import SvgIcon from "./SvgIcon.svelte";
   import { hrlToString } from "@holochain-open-dev/utils";
